Revision 2 of Ivth is mostly done.
This revision changes a lot of the underlying architecture. More
things are moved into the main memory to expose them to Ivth
operators.

The biggest change is to the 0br operator. Previously we followed
the forth convention of putting the memory address after the operation.
For instance "TEST 0br ADDR". There isn't a lot of reason to do
this and it makes other things more difficult so we've moved to a more
consistnt "TEST ADDR 0br" form.

There are also some smaller changes. For instance, literals take two
slots (-2 VALUE) so branch calculations must take this into account.
Once this is done I'll try to write a more full description.